// src/payload/payload-types.ts
export type CartItems = {
  product?: string | Product;
  quantity?: number;
  id?: string;
}[];
// ...
export interface Order {
  id: string;
  orderedBy?: string | User;
  stripePaymentIntentID?: string;
  total: number;
  items?: {
    product: string | Product;
    price?: number;
    quantity?: number;
    id?: string;
  }[];
  updatedAt: string;
  createdAt: string;
}
export interface User {
  id: string;
  name?: string;
  roles?: ('admin' | 'customer')[];
  purchases?: string[] | Product[];
  stripeCustomerID?: string;
  cart?: {
    items?: CartItems;
  };
  skipSync?: boolean;
  updatedAt: string;
  createdAt: string;
  email: string;
  resetPasswordToken?: string;
  resetPasswordExpiration?: string;
  salt?: string;
  hash?: string;
  loginAttempts?: number;
  lockUntil?: string;
  password: string;
}